Move MockMessageQueue
